pyproject.toml
[project]
name = "bfsi-bot-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "google-genai>=1.14.0",
    "langchain>=0.3.25",
    "langchain-anthropic>=0.3.12",
    "langchain-community>=0.3.23",
    "langchain-core>=0.3.59",
    "langchain-google-genai>=2.1.4",
    "langchain-openai>=0.3.16",
    "langgraph>=0.4.3",
    "llama-cloud-services>=0.6.21",
    "passlib>=1.7.4",
    "psycopg>=3.2.7",
    "pydantic[email]>=2.11.4",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "streamlit>=1.45.0",
    "uvicorn>=0.34.2",
]
